Update responsive-navbar.css
Updating responsive-navbar.css with additional functionality. There is now a hamburger icon for vertical navbars that will animate and turn into an x when the user opens the navbar. Code for sliding in navbars was also added, but commented out because it was not working correctly. Also refactored CSS and added a table of contents.